--- api/ExpressedRealms.Admin.API/AdminCharacterList/UpdatePlayer/UpdatePlayerEndpoint.cs
@@ -1,29 +1,0 @@
-using ExpressedRealms.Admin.UseCases.UpdatePlayer;
-using ExpressedRealms.Server.Shared;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Http.HttpResults;
-using Microsoft.AspNetCore.Mvc;
-
-namespace ExpressedRealms.Admin.API.AdminCharacterList.UpdatePlayer;
-
-public static class UpdatePlayerEndpoint
-{
-    public static async Task<Results<Ok, ValidationProblem, NotFound>> ExecuteAsync(
-        Guid characterId,
-        [FromBody] UpdatePlayerRequest request,
-        [FromServices] IUpdatePlayerUseCase useCase
-    )
-    {
-        var results = await useCase.ExecuteAsync(
-            new UpdatePlayerModel() { Id = characterId, PlayerNumber = request.PlayerNumber }
-        );
-
-        if (results.HasValidationError(out var validationProblem))
-            return validationProblem;
-        if (results.HasNotFound(out var notFound))
-            return notFound;
-        results.ThrowIfErrorNotHandled();
-
-        return TypedResults.Ok();
-    }
-}
