+++ api/ExpressedRealms.Admin.UseCases/GetPlayer/GetPlayerModelValidator.cs
@@ -1,4 +1,4 @@
-using ExpressedRealms.Characters.Repository.Players;
+using ExpressedRealms.Admin.Repository;
 using FluentValidation;
 using JetBrains.Annotations;
 
@@ -7,12 +7,12 @@ namespace ExpressedRealms.Admin.UseCases.GetPlayer;
 [UsedImplicitly]
 internal sealed class GetPlayerModelValidator : AbstractValidator<GetPlayerModel>
 {
-    public GetPlayerModelValidator(IPlayerRepository repository)
+    public GetPlayerModelValidator(IUsersRepository repository)
     {
         RuleFor(x => x.Id)
             .NotEmpty()
             .WithMessage("Id is required.")
-            .MustAsync(async (x, y) => await repository.PlayerExistsAsync(x))
-            .WithMessage("Character does not exist.");
+            .MustAsync(async (x, y) => await repository.UserExistsAsync(x.ToString()))
+            .WithMessage("User does not exist.");
     }
 }
