+++ api/ExpressedRealms.Characters.Repository/Players/IPlayerRepository.cs
@@ -1,4 +1,3 @@
-using ExpressedRealms.Characters.Repository.Players.Dtos;
 using ExpressedRealms.DB.UserProfile.PlayerDBModels.PlayerSetup;
 using ExpressedRealms.Shared;
 
@@ -10,7 +9,4 @@ public interface IPlayerRepository : IGenericRepository
     Task<bool> PlayerExistsAsync(Guid id);
     Task<string> GetUniqueLookupId();
     Task<Player> GetPlayerByCharacterId(int characterId);
-    Task<Player> GetPlayerByIdForEdit(Guid playerId);
-    Task<bool> PlayerNumberExists(int playerNumber);
-    Task<PlayerBasicInfoDto> GetPlayerBasicInfoAsync(Guid id);
 }
